+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - added: Done button is a part of SimpleTextInput
+- changed: Better send error message for insufficient gas
 - fixed: Glitch on transaction list scene when starting search
 - removed: 'X' button on modals