+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added: Done button is a part of SimpleTextInput
 - changed: Better send error message for insufficient gas
 - changed: Upgrade to React Native v0.73.6
+- fixed: Possible to see receive address for a wallet requiring activation when creating the wallet from search
 - fixed: Glitch on transaction list scene when starting search
 - removed: 'X' button on modals