+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - changed: "Split" wallet menu options consolidated into one
 - fixed: Possible to see receive address for a wallet requiring activation when creating the wallet from search
 - fixed: Glitch on transaction list scene when starting search
+- fixed: Properly set default buy/sell fiat amount for non-USD wallets
 - removed: 'X' button on modals
 
 ## 4.5.0